% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_diagnostic.R
\name{pai_diagnostic}
\alias{pai_diagnostic}
\title{PAI Diagnostic Tool}
\usage{
pai_diagnostic(
  pai_object = NULL,
  variables = NULL,
  bins = NULL,
  bin_cut = NULL
)
}
\arguments{
\item{pai_object}{A compiled PAI object obtained using pai_main().}

\item{variables}{A defined list of variables to analyze.}

\item{bins}{Numeric value of bins to assess predicted response across variance in the distribution of a variable; Default = 5}

\item{bin_cut}{Numeric value of cut point between bins; Default = 5}
}
\description{
PAI Diagnostic Tool
}
