% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_diagnostic.R
\name{pai_diagnostic_retrieval}
\alias{pai_diagnostic_retrieval}
\title{PAI Diagnostic Retrieval Function}
\usage{
pai_diagnostic_retrieval(
  output,
  diagnostic,
  type,
  variables = NULL,
  combine_plots = FALSE
)
}
\arguments{
\item{output}{Output object from `pai()`}

\item{diagnostic}{Diagnostic element to return. Options: `placebo`, `push`, or `bootstrap`}

\item{type}{Diagnostic subset to return. Options `data`, `figure`, or `distribution` (`bootstrap` ONLY)}

\item{variables}{Subset of variables to return (`figure` ONLY)}

\item{combine_plots}{Return `figure`(s) as combined plot using `cowplot` (CRAN). Defaults to `FALSE`}
}
\description{
PAI Diagnostic Retrieval Function
}
\examples{
Placebo Example
placebo <- pai_diagnostic_retrieval(output = output, diagnostic = 'placebo')
Push Example with combined plots and subset of var2 and var4
push <- pai_diagnostic_retrieval(output = output, diagnostic = 'push', type = 'figure', variables = c('var2', 'var4'), combine_plots = TRUE)
Bootstrap Example with variable-level plots
boostrap <- pai_diagnostic_retrieval(output = output, diagnostic = 'boostrap', type = 'figure', combine_plots = TRUE)
Boostrap Distribution Example
boostrap_distribution <- pai_diagnostic_retrieval(output = output, diagnostic = 'boostrap', type = 'distribution')
}
