% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_function.R
\name{pai_main}
\alias{pai_main}
\title{PAI Function}
\usage{
pai_main(
  data,
  outcome = NULL,
  predictors = NULL,
  factors = NULL,
  assign_factors = c(FALSE, 4),
  interactions = NULL,
  list_drop_vars = FALSE,
  drop_vars = NULL,
  ml = c(NA, NA, 10, 10),
  custom_tc = FALSE,
  seed = 1234
)
}
\arguments{
\item{data}{A structured dataset featuring a target variable of interest (y) along with independent variables(s) (x), providing a comprehensive framework for exploratory analysis and predictive modeling.}

\item{outcome}{Target variable of interest (y); Can be distributed binomialy or continuously.}

\item{predictors}{Independent variable(s); can declare directly or defaults to all if blank or NULL.}

\item{factors}{Declare factor terms found within declared predictors.}

\item{assign_factors}{Can assign factor status (TRUE/FALSE) and number of levels; defaults to FALSE and 4 levels.}

\item{interactions}{Declare interactive terms; Define each interaction within a vector as "Var1*Var2" & separated by a comma.}

\item{list_drop_vars}{Declare whether dropped terms should be grouped; define the grouped variables as an object in the Global Environment.}

\item{drop_vars}{Declare specific terms (or grouped variables) to drop.}

\item{ml}{(1) ML model to employ (see Caret Documentation); (2) Parallel computing cores; (3) Placebo Iterations; (4) K-Fold Cross Validations}

\item{custom_tc}{Declare custom train control parameters (see Caret Documentation for availble permutations); defaults to FALSE}

\item{seed}{Random seed generator; defaults to '1234'}
}
\description{
PAI Function
}
