% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_main.R
\name{pai}
\alias{pai}
\title{PAI Core Function}
\usage{
pai(
  data,
  model = NULL,
  outcome = NULL,
  predictors = NULL,
  interactions = NULL,
  drop_vars = NULL,
  cores = NULL,
  placebo_iterations = NULL,
  folds = NULL,
  train_split = 0.8,
  custom_tc = FALSE,
  assign_factors = 3,
  list_drop_vars = FALSE,
  seed = 1234
)
}
\arguments{
\item{data}{A structured dataset featuring a target variable of interest (y) and other predictors , providing a comprehensive framework for exploratory analysis and predictive modeling.}

\item{model}{Machine learning model to employ in protocol; see `caret` (CRAN) for more information on available models.}

\item{outcome}{Target variable of interest (y); Can be distributed binomialy or continuously.}

\item{predictors}{Covariate(s); can declare directly or defaults to all if blank or NULL.}

\item{interactions}{Interactive term(s); Define each interaction within a vector as "Var1*Var2" & separated by a comma.}

\item{drop_vars}{Specific terms from `predictors` and (or) `interactions` to drop during information omission protocol.}

\item{cores}{Number of computing cores to employ for parallel processing; Defaults to 1.}

\item{placebo_iterations}{Number of placebo iterations to employ; Defaults to 10.}

\item{folds}{Number of K-Folds to employ during `trainControl`; Defaults to 5}

\item{train_split}{Proportion of data for Train/Test split; Defaults to 0.8 (Indicating an 80/20 Train/Test Split).}

\item{custom_tc}{Custom `trainControl` parameters for unique model selection from `caret`; Defaults to common parameters for `parRF`, `adaboost`, and `rf` (i.e., tree-based methods).}

\item{assign_factors}{Integer representing number of levels to auto-assign factor terms; Defaults to 3.}

\item{list_drop_vars}{If TRUE, allows for the assignment of several variables to drop concurrently as indicated by objects stored in global environment.}

\item{seed}{Random Seed; Defaults to 1234}
}
\description{
PAI Core Function
}
