% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pai_diagnostic.R
\name{pai_diagnostic_retrieval}
\alias{pai_diagnostic_retrieval}
\title{PAI Diagnostic Retrieval Function}
\usage{
pai_diagnostic_retrieval(
  output,
  diagnostic,
  type,
  variables = NULL,
  combine_plots = FALSE,
  rename_y_title = NULL,
  rename_x_title = NULL,
  rename_x_labels = NULL,
  rename_y_labels = NULL,
  rename_title = NULL,
  rename_subtitle = NULL,
  rename_caption = NULL
)
}
\arguments{
\item{output}{Output object from `pai()`}

\item{diagnostic}{Diagnostic element to return. Options: `placebo`, `push`, `bootstrap`, or `summary`}

\item{type}{Diagnostic subset to return. Options `data`, `figure`, or `distribution` (`bootstrap` ONLY)}

\item{variables}{Subset of variables to return (`figure` ONLY)}

\item{combine_plots}{Return `figure`(s) as combined plot using `cowplot` (CRAN). Defaults to `FALSE`}
}
\description{
PAI Diagnostic Retrieval Function
}
\examples{
# placebo <- pai_diagnostic_retrieval(output = output, diagnostic = 'all')
# push <- pai_diagnostic_retrieval(output = output, diagnostic = 'push', type = 'figure', variables = c('var2', 'var4'), combine_plots = TRUE)
# bootstrap <- pai_diagnostic_retrieval(output = output, diagnostic = 'bootstrap', type = 'figure', combine_plots = TRUE)
# bootstrap_distribution <- pai_diagnostic_retrieval(output = output, diagnostic = 'bootstrap', type = 'distribution')
}
